handle exception on notification gateways by discovery to have safe registration process

+++ discovery-service/src/main/java/org/zowe/apiml/discovery/GatewayNotifier.java
@@ -9,21 +9,24 @@
  */
 package org.zowe.apiml.discovery;
 
-import org.zowe.apiml.message.core.MessageService;
-import org.zowe.apiml.message.log.ApimlLogger;
-import org.zowe.apiml.util.EurekaUtils;
 import com.netflix.appinfo.InstanceInfo;
 import com.netflix.discovery.shared.Application;
 import com.netflix.eureka.EurekaServerContext;
 import com.netflix.eureka.EurekaServerContextHolder;
 import com.netflix.eureka.registry.PeerAwareInstanceRegistry;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
+import org.zowe.apiml.message.core.MessageService;
+import org.zowe.apiml.message.log.ApimlLogger;
+import org.zowe.apiml.util.EurekaUtils;
 
 import java.util.List;
 
 @Component
+@Slf4j
 public class GatewayNotifier {
 
     private final ApimlLogger logger;
@@ -43,7 +46,7 @@ public class GatewayNotifier {
         return getServerContext().getRegistry();
     }
 
-    public void serviceUpdated(String serviceId) {
+    public void serviceUpdated(String serviceId, String instanceId) {
         final PeerAwareInstanceRegistry registry = getRegistry();
         final Application application = registry.getApplication("GATEWAY");
         if (application == null) {
@@ -54,11 +57,19 @@ public class GatewayNotifier {
         final List<InstanceInfo> gatewayInstances = application.getInstances();
 
         for (final InstanceInfo instanceInfo : gatewayInstances) {
+            // dont notify service itself, it is not required
+            if (StringUtils.equalsIgnoreCase(instanceId, instanceInfo.getInstanceId())) continue;
+
             final StringBuilder url = new StringBuilder();
             url.append(EurekaUtils.getUrl(instanceInfo)).append("/cache/services");
-            if (serviceId != null)
-                url.append('/').append(serviceId);
-            restTemplate.delete(url.toString());
+            if (serviceId != null) url.append('/').append(serviceId);
+
+            try {
+                restTemplate.delete(url.toString());
+            } catch (Exception e) {
+                log.debug("Cannot notify gateway {} about {}", url.toString(), instanceId, e);
+                logger.log("org.zowe.apiml.discovery.registration.gateway.notify", url.toString(), instanceId);
+            }
         }
     }
